requirements.txt
APScheduler==3.6.3
asgiref==3.8.1
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
idna==3.10
pillow==11.0.0
python-telegram-bot==13.15
pytz==2024.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
tornado==6.1
tzlocal==5.2
urllib3==2.2.3
